backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==3.0.10
Flask-SocketIO==5.6.0
PyJWT==2.6.0
gunicorn==21.2.0
boto3==1.42.23
